YOLOV5+Pseudo Labeling
Original submission used YOLOV5 model with Pseudo Labeling method, which used OOF-evaluation to search the best `score_threshold` for final prediction.  
Discarded this kernel due to the competition policy restrict the use of YOLOV5, so I finally chose the effnet and other models, which is under MIT license, to be my final submissions.